+++ packages/expo/ios/Fetch/NativeRequestCredentials.swift
@@ -1,0 +1,11 @@
+// Copyright 2015-present 650 Industries. All rights reserved.
+
+import ExpoModulesCore
+
+/**
+ Enum for RequestInit.credentials.
+ */
+internal enum NativeRequestCredentials: String, Enumerable {
+  case include
+  case omit
+}
