+++ packages/expo/src/winter/fetch/FetchErrors.ts
@@ -1,11 +1,11 @@
 export class NetworkFetchError extends Error {
-  constructor(message: string, cause?: unknown, stack?: string) {
+  constructor(message: string, { cause, stack }: { cause?: unknown; stack?: string } = {}) {
     super(`fetch failed: ${message}`);
     this.cause = cause;
     this.stack = stack;
   }
 
   static createFromError(error: Error): NetworkFetchError {
-    return new NetworkFetchError(error.message, error, error.stack);
+    return new NetworkFetchError(error.message, { cause: error.cause, stack: error.stack });
   }
 }
